Refactor folder structure, update link
