+++ example/lib/complex_example/main.dart
@@ -13,6 +13,17 @@ import 'hex_picker_page.dart';
 import 'alpha_picker_page.dart';
 import 'color_picker_page.dart';
 
+void main() => runApp(const MyApp());
+
+class MyApp extends StatelessWidget {
+  const MyApp({Key? key}) : super(key: key);
+
+  @override
+  Widget build(BuildContext context) {
+    return const MainPage();
+  }
+}
+
 class MainPage extends StatefulWidget {
   const MainPage({Key? key}) : super(key: key);
