+++ example/lib/complex_example/palette_hue_picker_page.dart
