+++ example/lib/complex_example/palette_picker_page.dart
