+++ example/lib/complex_example/palette_value_picker_page.dart
