+++ example/lib/complex_example/slider_picker_page.dart
