+++ example/lib/complex_example/swatches_picker_page.dart
