+++ example/lib/main.dart
@@ -3,7 +3,7 @@ import 'package:flutter/material.dart';
 import 'package:flutter_hsvcolor_picker/flutter_hsvcolor_picker.dart';
 
 /// For a more detailed example, head over to:
-/// https://github.com/fluttercandies/flutter_hsvcolor_picker/example/lib/complex_example_main.dart
+/// https://github.com/fluttercandies/flutter_hsvcolor_picker/example/lib/complex_example
 
 void main() async {
   runApp(const MyApp());
